pyproject.toml
[project]
name = "5bim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "paho-mqtt>=2.1.0",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
]
